requirements.txt
-i https://pypi.org/simple
absl-py==2.1.0; python_version >= '3.7'
aiofiles==23.2.1; python_version >= '3.7'
aiohttp==3.9.5; python_version >= '3.8'
aiosignal==1.3.1; python_version >= '3.7'
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.4.0; python_version >= '3.8'
asttokens==2.4.1
attrs==23.2.0; python_version >= '3.7'
certifi==2024.6.2; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
contourpy==1.2.1; python_version >= '3.9'
cycler==0.12.1; python_version >= '3.8'
decorator==5.1.1; python_version >= '3.5'
distro==1.9.0; python_version >= '3.6'
etils[epath,epy]==1.9.2; python_version >= '3.11'
eval-type-backport==0.2.0; python_version >= '3.8'
executing==2.0.1; python_version >= '3.5'
filelock==3.15.1; python_version >= '3.8'
fonttools==4.53.0; python_version >= '3.8'
frozenlist==1.4.1; python_version >= '3.8'
fsspec==2024.6.0
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.5; python_version >= '3.8'
httpx==0.27.0; python_version >= '3.8'
idna==3.7; python_version >= '3.5'
importlib-resources==6.4.0
ipython==8.25.0; python_version >= '3.10'
isodate==0.6.1
jedi==0.19.1; python_version >= '3.6'
jinja2==3.1.4; python_version >= '3.7'
jsonpath-rw==1.4.0
jsonpickle==3.2.1; python_version >= '3.7'
kiwisolver==1.4.5; python_version >= '3.7'
markdown-it-py==3.0.0; python_version >= '3.8'
markupsafe==2.1.5; python_version >= '3.7'
matplotlib==3.9.0; python_version >= '3.9'
matplotlib-inline==0.1.7; python_version >= '3.8'
mdurl==0.1.2; python_version >= '3.7'
mlcroissant==1.0.5
multidict==6.0.5; python_version >= '3.7'
networkx==3.3; python_version >= '3.10'
numpy==1.26.4; python_version >= '3.9'
openai==1.34.0; python_full_version >= '3.7.1'
packaging==24.1; python_version >= '3.8'
pandas==2.2.2; python_version >= '3.9'
pandas-stubs==2.2.2.240603; python_version >= '3.9'
parso==0.8.4; python_version >= '3.6'
pexpect==4.9.0; sys_platform != 'win32' and sys_platform != 'emscripten'
pillow==10.3.0; python_version >= '3.8'
ply==3.11
polars==0.20.31; python_version >= '3.8'
prompt-toolkit==3.0.47; python_full_version >= '3.7.0'
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0; python_version >= '3.8'
pydantic==2.7.4; python_version >= '3.8'
pydantic-core==2.18.4; python_version >= '3.8'
pygments==2.18.0; python_version >= '3.8'
pyparsing==3.1.2; python_full_version >= '3.6.8'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python-dotenv==1.0.1; python_version >= '3.8'
pytz==2024.1
pyvis==0.3.2; python_version >= '3.7'
rdflib==7.0.0; python_full_version >= '3.8.1' and python_full_version < '4.0.0'
regex==2024.5.15; python_version >= '3.8'
requests==2.32.3; python_version >= '3.8'
rich==13.7.1; python_full_version >= '3.7.0'
scipy==1.13.1; python_version >= '3.9'
seaborn==0.13.2; python_version >= '3.8'
shellingham==1.5.4; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sniffio==1.3.1; python_version >= '3.7'
stack-data==0.6.3
tabulate==0.9.0; python_version >= '3.7'
tenacity==8.3.0; python_version >= '3.8'
together==1.2.1; python_version >= '3.8' and python_version < '4.0'
tqdm==4.66.4; python_version >= '3.7'
traitlets==5.14.3; python_version >= '3.8'
typer==0.12.3; python_version >= '3.7'
types-pytz==2024.1.0.20240417; python_version >= '3.8'
typing-extensions==4.12.2; python_version >= '3.8'
tzdata==2024.1; python_version >= '2'
urllib3==2.2.1; python_version >= '3.8'
wcwidth==0.2.13
yarl==1.9.4; python_version >= '3.7'
zipp==3.19.2
